Get the direction of the language from a locale
Asking the direction from a locale, will try to get the script from the
locale, fallback on the language and return either `:left_to_right`,
`:right_to_left` or `:unknown_direction` (which should not happen often
but still).
Adds a `Kaur.Result.Map` module to deal with the weird result returned
from `Map.fetch`: a tuple on success but an atom on error.
With that, a convenience function `fetch_with_fallback` that tries a
series of keys before returning an error.